Fix submit incident
