+++ backend/docker-compose.dev.yml
@@ -19,7 +19,7 @@ services:
   postgres:
     image: postgres:17
     ports:
-      - '5454:5432'
+      - '5432:5432'
     env_file:
       - .env
     volumes:
